Refactor: Final clean state with BigDecimal precision and TDD tests

+++ src/main/java/com/incubyte/salary/dto/EmployeeDto.java
@@ -1,6 +1,7 @@
 package com.incubyte.salary.dto;
 
 import lombok.*;
+import java.math.BigDecimal; // Add this import
 
 @Data
 @NoArgsConstructor
@@ -10,5 +11,5 @@ public class EmployeeDto {
     private String fullName;
     private String jobTitle;
     private String country;
-    private Double salary;
+    private BigDecimal salary; // Changed from Double to BigDecimal
 }
